requirements.txt
django==3.2.0
django-environ
djangorestframework==3.13.0
django-oauth-toolkit==2.0.0
psycopg2==2.9.3
django-cors-middleware==1.5.0
django-debug-toolbar==3.4.0
requests